requirements.txt
pyspark==3.0.1
jupyter==1.0.0
pandas==1.3.5
pandas-datareader==0.10.0
tensorflow-hub
keras
numpy
pytest
matplotlib
scikit-learn
pyyaml
ipykernel
kafka-python
findspark
torchtext==0.6.0
spacy
torch==1.8.0
GoogleNews==1.4.1
newspaper3k==0.2.8
pytz
searchtweets==1.7.6
xgboost==1.6.0
watermark==2.0.2
transformers==4.29.2
seaborn==0.11.0
vaderSentiment
pandas_datareader
pyarrow==3.0.0
dask[complete]==2.30.0
slackclient==2.2.0
yfinance==0.1.54
retry==0.9.2
schedule==1.0.0
shap==0.40.0
catboost==1.0.6
polygon-api-client
finnhub-python
retry-decorator==1.1.1
pyaml
spacy==3.5.3
scikit_learn==1.0.1
matplotlib==3.6.3
alpaca-py==0.8.2
mplchart==0.0.2
diskcache==5.6.3